fix(cloudflare/workers): rely on rolldown-plugin for esmExternalRequire
Bump cloudflare-tools so esmExternalRequirePlugin is registered at the
plugin top level (alchemy-run/cloudflare-tools#77), drop the alchemy-side
duplicate that papered over the options-hook identity mismatch, and fold
the #880 deploy regression into WorkerBundle.test.ts.
